Moved TextField to its own component
